feat: add StreamHTTPRecorder and tests for streamable HTTP responses
